Moved methods regarding traveable zones from Zone -> InterZoneTravel
