+++ src/GameObjects/Worlds/Zone.java
@@ -165,42 +165,6 @@ public class Zone {
 
     }
 
-    public Zone displayTraveableZones(PlayerCharacter pc) {
-
-        Utility.clearConsole();
-
-        int index = 1;
-
-        System.out.println("Travelable Zones:");
-        Iterator<Zone> zones = traveableZones.iterator(); 
-        while (zones.hasNext()) {
-            System.out.println(index + ". " + zones.next().getName());
-            index++;
-        }
-
-        System.out.println("Enter the number of the zone you want to travel to: ");
-        int choice = Utility.checkIfNumber(sc);
-        // check if choice is valid, make array of traveable zones to be able to index it for selection
-        if (choice > 0 && choice <= traveableZones.size()) {
-            Zone[] zonesArray = traveableZones.toArray(Zone[]::new); 
-            Zone selectedZone = zonesArray[choice - 1]; 
-             // check if player is already in the selected zone 
-            if (pc.getCurrentZone().equals(selectedZone)) {
-                Utility.clearConsole();
-                Utility.slowPrint("You return to the " + selectedZone.getName());
-                return pc.getCurrentZone(); 
-            }
-            Utility.clearConsole();
-            Utility.slowPrint("You travel to the " + selectedZone.getName());
-            return selectedZone;
-        // error handling for invalid choice
-        } else { 
-            System.out.println("Invalid choice. Please try again. Or: ");
-            Utility.promptEnterKey(sc);
-            // Recursively call the method again for a valid choice
-            return displayTraveableZones(pc); 
-        }
-    }
     // Just displays the current zone and its description + clear status
     public void displayCurrentZone(PlayerCharacter pc) { 
 
@@ -209,28 +173,6 @@ public class Zone {
                 + " Zone cleared: " + pc.getCurrentZone().getZoneCleared());
         Utility.promptEnterKey(sc);
     }
-     // Adds traveable zones to a zone.
-    public static void addTraveableZone(Zone zone, ZoneType zonetype) { 
-        if (zone != null) {
-            zone.getTraveableZones().add(ZoneManager.getZone(zonetype));
-        }
-    }
-    // Adds traveable zones based on cleared zones.
-    public void checkTraveableZones(PlayerCharacter pc) { 
-
-        if ((getZoneType() == ZoneType.FOREST && getZoneCleared())
-                || (getZoneType() == ZoneType.SWAMP && getZoneCleared())) {
-            addTraveableZone(this, ZoneType.CAVE);
-            addTraveableZone(ZoneManager.getZone(ZoneType.TAVERN), ZoneType.CAVE);
-        }
-
-        if (getZoneType() == ZoneType.CAVE && getZoneCleared()) {
-            addTraveableZone(this, ZoneType.BASEMENT);
-            addTraveableZone(ZoneManager.getZone(ZoneType.TAVERN), ZoneType.BASEMENT);
-            addTraveableZone(ZoneManager.getZone(ZoneType.FOREST), ZoneType.BASEMENT);
-            addTraveableZone(ZoneManager.getZone(ZoneType.SWAMP), ZoneType.BASEMENT);
-        }
-
-    }
+     
 
 }
